refactor(notifications): move notification helper functions to a dedicated helpers module
